--- execution-plane/arcon/arcon_backend/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "arcon_backend"
-version = "0.1.0"
-authors = ["Max Meldrum <mmeldrum@kth.se>"]
-edition = "2018"
-readme = "README.md"
-license = "BSD-3-Clause"
-repository = "https://github.com/cda-group/arcon"
-homepage = "https://github.com/cda-group/arcon"
-description = """
-State Backends for Arcon
-"""
-
-[features]
-default = []
-
-[dependencies]
-arcon_error = { path = "../arcon_error", version = "0.1.0"}
-
-[dependencies.rocksdb]
-version = "0.12.1"
-default-features = false
-features = ["lz4"]
-optional = true
-
-[dev-dependencies]
-tempfile = "3"
