--- execution-plane/arcon/arcon_backend/README.md
@@ -1,7 +1,0 @@
-[![Build Status](https://dev.azure.com/arcon-cda/arcon/_apis/build/status/cda-group.arcon?branchName=master)](https://dev.azure.com/arcon-cda/arcon/_build/latest?definitionId=1&branchName=master)
-[![Cargo](https://img.shields.io/badge/crates.io-v0.1.0-orange)](https://crates.io/crates/arcon_backend)
-[![License](https://img.shields.io/badge/License-BSD%203--Clause-blue)](https://github.com/cda-group/arcon)
-
-# arcon backend
-
-This crate is not intended to be used directly. It is a part of the [Arcon project](https://github.com/cda-group/arcon)
