+++ execution-plane/arcon/arcon_messages/Cargo.toml
@@ -17,7 +17,7 @@ proto = ["protobuf", "protoc-rust"]
 capnproto = ["capnp", "derivative", "capnpc"]
 
 [dependencies]
-kompact = { git = "https://github.com/Max-Meldrum/kompact.git", branch = "dedicated_investigation" }
+kompact = { git = "https://github.com/kompics/kompact.git", branch = "master" }
 protobuf = { version = "2.0", optional = true }
 capnp = { version = "0.9.4", optional = true }
 derivative = { version = "1.0.2", optional = true }
