+++ execution-plane/arcon/src/state_backend/mod.rs
@@ -2,14 +2,14 @@
 extern crate tempfile;
 
 pub mod in_memory;
-#[cfg(feature = "rocksdb")]
+#[cfg(feature = "arcon_rocksdb")]
 pub mod rocksdb;
 
 use arcon_error::ArconResult;
 
 /// Trait required for all state backend implementations in Arcon
 pub trait StateBackend: Send + Sync {
-    fn create(name: &str) -> Self
+    fn create(name: &str) -> ArconResult<Self>
     where
         Self: Sized;
     fn put(&mut self, key: &[u8], value: &[u8]) -> ArconResult<()>;
